Add reusable component CommentArea and refactor PostPreview
