+++ src/assets/styles/cmps/CommentArea.scss
@@ -1,0 +1,67 @@
+.comment-count {
+    font-size: rem(14px);
+    line-height: rem(18px);
+    // font-weight: bold;
+    margin-top: 14px;
+    color: ($secondary-color-1);
+}
+
+.comment-area {
+    margin-top: 5px;
+    margin-left: -1px;
+    position: relative;
+    display: flex;
+    gap: 8px;
+    align-items: center;
+
+    // border: 4px solid red;
+    height: 22px;
+    max-height: 50px;
+
+    textarea {
+       // height: rem(22px);
+       // max-height: calc(rem(22px) * 4);
+       height: 100%;
+       overflow: auto;
+       flex: 1;
+       border: none;
+       background-color: $primary-bgc;
+       color: ($primary-color-1);
+       resize: none;
+       font-size: rem(14px);
+       line-height: rem(18px);
+
+       &::placeholder {
+           color: $secondary-color-1
+       }
+
+       &:focus {
+           outline: none;
+           border: none;
+       }
+       // border: 4px solid blue;
+    }
+
+    .submit-comment {
+          height: auto;
+          background-color: $primary-bgc;
+          color: $btn-secondary-bgc;
+          font-weight: bold;
+          cursor: pointer;
+
+          &:hover {
+              color: $primary-color-1;
+          }
+          // border: 4px solid greenyellow;
+    }
+
+    .emoji {
+       cursor: pointer;
+    }
+
+   .emoji-picker {
+       position: absolute;
+       bottom: 50px;
+       right: 10px;
+   }
+ }
