+++ src/assets/styles/main.scss
@@ -21,6 +21,7 @@
 @import "cmps/AppFooter";
 @import "cmps/ActionAlert";
 @import "cmps/ActionButtons";
+@import "cmps/CommentArea.scss";
 @import "cmps/CreatePost.scss";
 @import "cmps/SideBarNavigation.scss";
 @import "cmps/PostPreview.scss";
